+++ sentry-android-ndk/src/main/java/io/sentry/android/ndk/SentryNdk.java
@@ -73,8 +73,8 @@ public final class SentryNdk {
           ndkOptions.setTracesSampleRate(tracesSampleRate.floatValue());
         }
 
-        ndkOptions.setEnableAppHangTracking(options.isEnableAppHangTracking());
-        ndkOptions.setAppHangTimeoutMillis(options.getAppHangTimeoutIntervalMillis());
+        ndkOptions.setEnableAppHangTracking(options.isEnableNdkAppHangTracking());
+        ndkOptions.setAppHangTimeoutMillis(options.getNdkAppHangTimeoutIntervalMillis());
 
         //noinspection UnstableApiUsage
         io.sentry.ndk.SentryNdk.init(ndkOptions);